+++ jest.config.mjs
@@ -3,7 +3,7 @@ import { createDefaultEsmPreset } from 'ts-jest';
 export default {
   ...createDefaultEsmPreset({ tsconfig: 'tsconfig.test.json' }),
   testEnvironment: 'node',
-  testMatch: ['**/*.test.ts'],
+  testMatch: ['**/*.test.ts', '**/*.test.js'],
   testPathIgnorePatterns: ['<rootDir>/dist/'],
   moduleNameMapper: {
     '^(\\.{1,2}/.*)\\.js$': '$1',